backend/requirements.txt
fastapi==0.104.1
pydantic>=2.4.2
pydantic[email]>=2.4.2
uvicorn==0.23.2
httpx==0.25.0
jieba==0.42.1
pygtrie==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
google-auth==2.22.0
google-auth-oauthlib==1.0.0
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
PyJWT==2.8.0
starlette==0.27.0
typing-extensions>=4.8.0
slowapi==0.1.8
limits==3.7.0
certifi>=2024.2.2